requirements.txt
easyocr==1.7.1
google-generativeai
qwen-vl-utils
transformers
flash-attn
scipy
yacs
datasets
transformers
SentencePiece
nltk
torch
accelerate
bitsandbytes
unidecode
flask
flask_cors
pillow